backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.109.0
google-auth==2.48.0
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.10.0
prisma==0.11.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
google-cloud-storage==2.18.2
# JWT: use PyJWT only; do not install the PyPI package named "jwt" (it shadows this one).
PyJWT==2.9.0
jwcrypto==1.5.0
pydantic_core==2.41.5
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
starlette==0.35.1
tenacity==9.1.2
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
spacy==3.8.11
markdown==3.10.1
ag-ui-protocol
# Multi-agent workflow; requires google-adk >= 1.0 (avoid legacy 0.0.1 which lacks google.adk module).
google-adk[a2a]>=1.0
reportlab==4.2.5
weasyprint==62.3
playwright==1.48.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-multipart==0.0.22
